+++ src/jdepend/framework/ClassContainers.java
@@ -36,4 +36,25 @@ class ClassContainers extends ArrayList<ClassContainer> {
 
         return files;
     }
+
+    boolean existsWithExtension(File file, String extension) {
+        return file.isFile() &&
+                file.getName().toLowerCase().endsWith(extension);
+    }
+
+    boolean isJar(File file) {
+        return existsWithExtension(file, ".jar");
+    }
+
+    boolean isZip(File file) {
+        return existsWithExtension(file, ".zip");
+    }
+
+    boolean isWar(File file) {
+        return existsWithExtension(file, ".war");
+    }
+
+    public boolean isValidContainer(File file) {
+        return isJar(file) || isZip(file) || isWar(file);
+    }
 }
